revert some of the larger scale reorganizations
