+++ .gitignore
@@ -1,2 +1,4 @@
 *pycache*
 *results*
+*data_ct*
+*groundtruth_ct*
