+++ src/load.py
@@ -353,47 +353,3 @@ def clean_merged(merged,setname):
     #explicitly each time it occurs in the dataset (because of the FSS)
     return merged.drop(labels='Count',axis='columns')
 
-########################################
-# Create imgtrain Overall Output Files #----------------------------------------
-########################################
-def combine_imgtrain_files(term_search_dir):
-    """Combine imgtrain_notetrain, imgtrain_notetest, and imgtrain_extra output
-    files.
-    This function is not for initial data loading. This function is called
-    in certain circumstances after SARLE has totally finished running
-    in order to aggregate certain output files."""
-    #Aggregate all training labels (location x disease) and save
-    imgtrain_notetrain = pickle.load(open(os.path.join(term_search_dir, 'imgtrain_notetrain_BinaryLabels.pkl'), 'rb'))
-    imgtrain_notetest = pickle.load(open(os.path.join(term_search_dir, 'imgtrain_notetest_BinaryLabels.pkl'), 'rb'))
-    imgtrain_extra = pickle.load(open(os.path.join(term_search_dir, 'imgtrain_extra_BinaryLabels.pkl'), 'rb'))
-    out_bin = {}
-    out_bin.update(imgtrain_notetrain)
-    out_bin.update(imgtrain_notetest)
-    out_bin.update(imgtrain_extra)
-    pickle.dump(out_bin, open(os.path.join(term_search_dir, 'imgtrain_BinaryLabels.pkl'),'wb'))
-    
-    #Aggregate disease_out (disease binary labels) and save
-    train_disease_out = (pd.concat([pd.read_csv(os.path.join(term_search_dir, 'imgtrain_notetrain_DiseaseBinaryLabels.csv'),
-                                       header = 0, index_col = 0),
-                           pd.read_csv(os.path.join(term_search_dir, 'imgtrain_notetest_DiseaseBinaryLabels.csv'),
-                                       header = 0, index_col = 0),
-                           pd.read_csv(os.path.join(term_search_dir, 'imgtrain_extra_DiseaseBinaryLabels.csv'),
-                                       header = 0, index_col = 0)],axis=0))
-    train_disease_out.to_csv(os.path.join(term_search_dir, 'imgtrain_DiseaseBinaryLabels.csv'))
-    
-    #Missingness
-    train_missing = (pd.concat([pd.read_csv(os.path.join(term_search_dir, 'imgtrain_notetrain_Missingness.csv'),
-                                       header = 0, index_col = 0),
-                           pd.read_csv(os.path.join(term_search_dir, 'imgtrain_notetest_Missingness.csv'),
-                                       header = 0, index_col = 0),
-                           pd.read_csv(os.path.join(term_search_dir, 'imgtrain_extra_Missingness.csv'),
-                                       header = 0, index_col = 0)],axis=0))
-    train_missing.to_csv(os.path.join(term_search_dir, 'imgtrain_Missingness.csv'))
-    
-    #Sanity checks
-    #columns ['MRN', 'Accession', 'Set_Assigned', 'Set_Should_Be','Subset_Assigned']
-    all_ids, available_accs = load_all_ids_and_accs()
-    train_ids = all_ids[all_ids['Set_Assigned'].isin(['imgtrain_extra','imgtrain_notetrain','imgtrain_notetest'])]['Accession'].values.tolist()
-    train_ids = [x for x in train_ids if x in available_accs]
-    assert set(train_ids)==set(list(out_bin.keys()))
-
